added image exports and refactored code
