+++ src/images/index.js
@@ -1,0 +1,23 @@
+export { default as benxene } from './benxene.png';
+
+export { default as female1 } from './female-1.png';
+export { default as female2 } from './female-2.png';
+export { default as female3 } from './female-3.png';
+
+export { default as male1 } from './male-1.png';
+export { default as male2 } from './male-2.png';
+export { default as male3 } from './male-3.png';
+
+export { default as man } from './man.png';
+
+export { default as bookmark } from './bookmark.svg';
+export { default as devStats } from './dev-github.svg';
+export { default as hero } from './hero.svg';
+export { default as interview } from './interview.svg';
+export { default as link } from './link.svg';
+export { default as share } from './share.svg';
+export { default as meeting } from './meeting.svg';
+export { default as team } from './team.svg';
+
+export { default as logoColored } from './logo-colored.svg';
+export { default as logoWhite } from './logo-white.png';
