+++ src/pages/home/Members.js
@@ -1,16 +1,16 @@
 import styled from 'styled-components';
 import { Card, Container, Heading } from '../../components';
 
-import Bheem from '../../images/man.png';
+import { man } from '../../images';
 
 export default function Members() {
 	return (
 		<Container>
 			<Heading>Our core members</Heading>
 			<CoreBox>
-				<Card imgSrc={Bheem} name='Bheem' position='Vice President' />
-				<Card imgSrc={Bheem} name='Bheem' position='President' isSpecial />
-				<Card imgSrc={Bheem} name='Bheem' position='Tech Lead' />
+				<Card imgSrc={man} name='Bheem' position='Vice President' />
+				<Card imgSrc={man} name='Bheem' position='President' isSpecial />
+				<Card imgSrc={man} name='Bheem' position='Tech Lead' />
 			</CoreBox>
 		</Container>
 	);
